+++ cmake/AstarteMQTTTransport.cmake
@@ -21,6 +21,8 @@ function(astarte_sdk_configure_mqtt_dependencies)
         find_package(PahoMqttCpp REQUIRED)
         find_package(cpr REQUIRED)
         find_package(nlohmann_json REQUIRED)
+        find_package(MbedTLS REQUIRED)
+
         if(NOT TARGET ada::ada)
             find_package(ada REQUIRED)
         endif()
@@ -54,6 +56,20 @@ function(astarte_sdk_configure_mqtt_dependencies)
         set(URL_GIT_TAG v3.2.4)
         FetchContent_Declare(ada GIT_REPOSITORY ${URL_GIT_REPOSITORY} GIT_TAG ${URL_GIT_TAG})
         FetchContent_MakeAvailable(ada)
+
+        # Cryptographic library
+        set(MBEDTLS_GIT_REPOSITORY https://github.com/Mbed-TLS/mbedtls.git)
+        set(MBEDTLS_GIT_TAG v4.0.0)
+        FetchContent_Declare(
+            MbedTLS
+            GIT_REPOSITORY ${MBEDTLS_GIT_REPOSITORY}
+            GIT_TAG ${MBEDTLS_GIT_TAG}
+        )
+        # Disable programs and tests to keep the build fast and minimal.
+        set(ENABLE_TESTING OFF CACHE BOOL "Disable Mbed TLS tests")
+        set(ENABLE_PROGRAMS OFF CACHE BOOL "Disable Mbed TLS example programs")
+        set(CONFIG_MBEDTLS_PSA_CRYPTO_C ON CACHE BOOL "Enable PSA in Mbed TLS")
+        FetchContent_MakeAvailable(MbedTLS)
     endif()
 endfunction()
 
@@ -83,8 +99,14 @@ function(
         "src/mqtt/device_mqtt_impl.cpp"
         "src/mqtt/pairing.cpp"
         "src/mqtt/config.cpp"
+        "src/mqtt/crypto.cpp"
     )
-    list(APPEND ${ASTARTE_MQTT_PRIVATE_HEADERS} "private/mqtt/device_mqtt_impl.hpp")
+    list(
+        APPEND
+        ${ASTARTE_MQTT_PRIVATE_HEADERS}
+        "private/mqtt/device_mqtt_impl.hpp"
+        "private/mqtt/crypto.hpp"
+    )
     set(${ASTARTE_MQTT_PUBLIC_HEADERS} ${${ASTARTE_MQTT_PUBLIC_HEADERS}} PARENT_SCOPE)
     set(${ASTARTE_MQTT_SOURCES} ${${ASTARTE_MQTT_SOURCES}} PARENT_SCOPE)
     set(${ASTARTE_MQTT_PRIVATE_HEADERS} ${${ASTARTE_MQTT_PRIVATE_HEADERS}} PARENT_SCOPE)
@@ -104,6 +126,8 @@ function(astarte_sdk_add_mqtt_transport)
         astarte_device_sdk
         PRIVATE cpr::cpr
         PRIVATE nlohmann_json::nlohmann_json
+        PUBLIC MbedTLS::mbedtls
+        PUBLIC MbedTLS::mbedx509
         PUBLIC ada::ada
     )
 endfunction()