+++ conanfile.py
@@ -38,6 +38,7 @@ class Pkg(ConanFile):
             self.requires("cpr/1.11.2")
             self.requires("nlohmann_json/3.12.0")
             self.requires("ada/3.2.4", transitive_headers=True)
+            self.requires("mbedtls/3.6.5")
         self.requires("spdlog/1.15.3", options={"use_std_fmt": "True"}, transitive_headers=True, transitive_libs=True)
 
     def build_requirements(self):