+++ src/mqtt/pairing.cpp
@@ -14,6 +14,7 @@
 #include <string_view>
 
 #include "ada.h"
+#include "mqtt/crypto.hpp"
 
 using json = nlohmann::json;
 
@@ -99,6 +100,87 @@ auto PairingApi::get_broker_url(std::string_view credential_secret, int timeout_
   try {
     json response_json = json::parse(res.text);
     return response_json.at("data").at("protocols").at("astarte_mqtt_v1").at("broker_url");
+  } catch (const json::exception& e) {
+    throw JsonAccessErrorException(
+        std::format("Failed to parse JSON: {}. Body: {}", e.what(), res.text));
+  }
+}
+
+auto PairingApi::get_device_cert(std::string_view credential_secret, int timeout_ms) const
+    -> std::string {
+  auto request_url = pairing_url;
+  std::string pathname = std::format("{}/v1/{}/devices/{}/protocols/astarte_mqtt_v1/credentials",
+                                     request_url.get_pathname(), realm, device_id);
+  request_url.set_pathname(pathname);
+  spdlog::debug("request url: {}", request_url.get_href());
+
+  cpr::Header header{{"Content-Type", "application/json"},
+                     {"Authorization", std::format("Bearer {}", credential_secret)}};
+
+  auto priv_key = PsaKey();
+  priv_key.generate();
+  auto device_csr = Crypto::create_csr(std::move(priv_key));
+
+  json body;
+  body["data"] = {{"csr", device_csr}};
+  spdlog::debug("request body: {}", body.dump());
+
+  cpr::Response res = cpr::Post(cpr::Url{request_url.get_href()}, header, cpr::Body{body.dump()},
+                                cpr::Timeout{timeout_ms});
+
+  if (res.error) {
+    throw DeviceRegistrationException(std::format(
+        "Failed to retrieve Astarte device certificate. CPR error: {}", res.error.message));
+  }
+
+  if (!is_successful(res.status_code)) {
+    throw DeviceRegistrationException(std::format(
+        "Failed to retrieve Astarte device certificate. HTTP status code: {}, Reason: {}",
+        res.status_code, res.text));
+  }
+
+  try {
+    json response_json = json::parse(res.text);
+    return response_json.at("data").at("client_crt");
+  } catch (const json::exception& e) {
+    throw JsonAccessErrorException(
+        std::format("Failed to parse JSON: {}. Body: {}", e.what(), res.text));
+  }
+}
+
+auto PairingApi::device_cert_valid(std::string_view certificate, std::string_view credential_secret,
+                                   int timeout_ms) const -> bool {
+  auto request_url = pairing_url;
+  std::string pathname =
+      std::format("{}/v1/{}/devices/{}/protocols/astarte_mqtt_v1/credentials/verify",
+                  request_url.get_pathname(), realm, device_id);
+  request_url.set_pathname(pathname);
+  spdlog::debug("request url: {}", request_url.get_href());
+
+  cpr::Header header{{"Content-Type", "application/json"},
+                     {"Authorization", std::format("Bearer {}", credential_secret)}};
+
+  json body;
+  body["data"] = {{"client_crt", certificate}};
+  spdlog::debug("request body: {}", body.dump());
+
+  cpr::Response res = cpr::Post(cpr::Url{request_url.get_href()}, header, cpr::Body{body.dump()},
+                                cpr::Timeout{timeout_ms});
+
+  if (res.error) {
+    throw DeviceRegistrationException(std::format(
+        "Failed to check Astarte device certificate validity. CPR error: {}", res.error.message));
+  }
+
+  if (!is_successful(res.status_code)) {
+    throw DeviceRegistrationException(std::format(
+        "Failed to check Astarte device certificate validity. HTTP status code: {}, Reason: {}",
+        res.status_code, res.text));
+  }
+
+  try {
+    json response_json = json::parse(res.text);
+    return response_json.at("data").at("valid");
   } catch (const json::exception& e) {
     throw JsonAccessErrorException(
         std::format("Failed to parse JSON: {}. Body: {}", e.what(), res.text));